Implement the first 3 smells for the Next button.
It would be misleading to report the detailed accuracy at this point; the trainer is assuming a lot of samples (ones with non-<button> buttons) are negative and declaring spurious success.

A rough pagewise training accuracy is something like 80%, but we're getting a lot of help on the negative samples by simply not extracting <input type=submit>s at all.

I'm also sure I'm overfitting, since I'm not using a validation set for this run.