Add <input> elements to consideration. 94% per-tag validation accuracy. Not bad for starters!
I had to give up on filtering input vs. button on the LHS in the absence of conserveScore(); that led to nulls in the vectors for rules that didn't apply for those nodes. So, instead, I moved the when() conditions onto the RHS, effectively transforming them to mean "matches the when() and whatever signal you were going for in the first place".

  Training accuracy per tag:  0.95789    95% CI: (0.92934, 0.98645)
Validation accuracy per tag:  0.94318    95% CI: (0.90898, 0.97738)
  Training accuracy per page: 0.98333    95% CI: (0.95094, 1.00000)
Validation accuracy per page: 0.95312    95% CI: (0.90134, 1.00000)