Divide samples into initial training, validation, and testing sets.
We did this by temporarily unioning positive and negative samples and using fathom-pick (followed by redoing it with `git mv`, which was a pain). But then we made some fine adjustments to make sure each set received some of each language where possible. We also ensured that Vlad's initial 31 samples, which were not drawn from the Trexa list, were fairly proportionally distributed.

Initial sets are 66 training, 45 validation, 43 testing, keeping some samples in reserve in case we need more (likely for training or if we have to burn the validation set). More samples are being gathered as well.