Adds a rule looking for remember me checkboxes.
Surprisingly, the rule does very little even though there are many candidates that have them. The reason is EN/RU_161. It uses the same form for both the login and registration form. We have a hard time with this one and the optimizer doesn't want to punish it even more.

`fathom-train training.json -a validation.json -s -i 10000 -l 0.001`

The old results are lower than the last rule PR #55 because #64 changed a training sample and that had a surprisingly large influence on the validation results.

New results:
```
  Training accuracy per tag:  0.98286    95% CI: (0.97479, 0.99094)
                        FPR:  0.00524    95% CI: (0.00012, 0.01035)
                        FNR:  0.05702    95% CI: (0.02692, 0.08712)
                  Precision:  0.98174    Recall: 0.94298
                   F1 Score:  0.96197

Validation accuracy per tag:  0.96367    95% CI: (0.94764, 0.97971)
                        FPR:  0.01703    95% CI: (0.00452, 0.02954)
                        FNR:  0.10714    95% CI: (0.04986, 0.16443)
                  Precision:  0.93458    Recall: 0.89286
                   F1 Score:  0.91324
```

Old results:
```
  Training accuracy per tag:  0.98286    95% CI: (0.97479, 0.99094)
                        FPR:  0.00524    95% CI: (0.00012, 0.01035)
                        FNR:  0.05702    95% CI: (0.02692, 0.08712)
                  Precision:  0.98174    Recall: 0.94298
                   F1 Score:  0.96197

Validation accuracy per tag:  0.96183    95% CI: (0.94543, 0.97824)
                        FPR:  0.01942    95% CI: (0.00609, 0.03274)
                        FNR:  0.10714    95% CI: (0.04986, 0.16443)
                  Precision:  0.92593    Recall: 0.89286
                   F1 Score:  0.90909
```

For fun I removed the EN/RU_161 samples and re-vectorized and trained. The coefficients for these new rules go up to -2.2 each implying it is a high signal rule, but the output metrics are the same. This is because all of the remaining FPs in the training set (and likely the validation set) don't have remember me checkboxes or aren't in forms.